+++ src/components/statsIncrease/VitIncrease.jsx
@@ -24,22 +24,22 @@ const VitIncrease = () => {
   };
 
   const aux = async (statName, stat, setStat) => {
-    const champ = await statusUpdate(statName, parseInt(stat), id);
+    const champ = await statusUpdate(statName, parseFloat(stat), id);
     dispatch(selectChampion(champ));
     setStat("");
   };
 
   return (
-    <div className='container-stats'>
-      <h4 className='h4-stats text-white/80 bg-blue-500'>Vitalidade</h4>
-      <div className=''>
-        <div className='mb-3'>
+    <div className="container-stats">
+      <h4 className="h4-stats text-white/80 bg-blue-500">Vitalidade</h4>
+      <div className="">
+        <div className="mb-3">
           <input
-            className='input-stats'
-            id='meals-increase'
-            placeholder='Refeições Saudáveis'
-            type='number'
-            name='meals'
+            className="input-stats"
+            id="meals-increase"
+            placeholder="Refeições Saudáveis"
+            type="number"
+            name="meals"
             value={meals}
             onChange={({ target: { value } }) =>
               handleChange(value, 0, 8, setMeals)
@@ -47,8 +47,8 @@ const VitIncrease = () => {
           />
 
           <button
-            className='btn-stats bg-blue-500 hover:bg-blue-400'
-            type='button'
+            className="btn-stats bg-blue-500 hover:bg-blue-400"
+            type="button"
             onClick={() => aux("meals", meals, setMeals)}
           >
             Enviar
@@ -56,14 +56,14 @@ const VitIncrease = () => {
         </div>
       </div>
 
-      <div className=''>
-        <div className='mb-3'>
+      <div className="">
+        <div className="mb-3">
           <input
-            className='input-stats'
-            id='drinks-increase'
-            placeholder='Litros de Água'
-            type='number'
-            name='drinks'
+            className="input-stats"
+            id="drinks-increase"
+            placeholder="Litros de Água"
+            type="number"
+            name="drinks"
             value={drinks}
             onChange={({ target: { value } }) =>
               handleChange(value, 0, 5, setDrinks)
@@ -71,8 +71,8 @@ const VitIncrease = () => {
           />
 
           <button
-            className='btn-stats bg-blue-500 hover:bg-blue-400'
-            type='button'
+            className="btn-stats bg-blue-500 hover:bg-blue-400"
+            type="button"
             onClick={() => aux("drinks", drinks, setDrinks)}
           >
             Enviar
@@ -80,14 +80,14 @@ const VitIncrease = () => {
         </div>
       </div>
 
-      <div className=''>
-        <div className='mb-3'>
+      <div className="">
+        <div className="mb-3">
           <input
-            className='input-stats'
-            id='sleep-increase'
-            placeholder='Horas de Sono'
-            type='number'
-            name='sleep'
+            className="input-stats"
+            id="sleep-increase"
+            placeholder="Horas de Sono"
+            type="number"
+            name="sleep"
             value={sleep}
             onChange={({ target: { value } }) =>
               handleChange(value, 0, 12, setSleep)
@@ -95,8 +95,8 @@ const VitIncrease = () => {
           />
 
           <button
-            className='btn-stats bg-blue-500 hover:bg-blue-400'
-            type='button'
+            className="btn-stats bg-blue-500 hover:bg-blue-400"
+            type="button"
             onClick={() => aux("sleep", sleep, setSleep)}
           >
             Enviar
